pyproject.toml
[tool.poetry]
name = "static-conversations"
version = "0.1.0"
description = ""
authors = ["Nick Anderegg <nick@anderegg.io>"]

packages = [
    { include = "convo" },
]

[tool.poetry.dependencies]
python = "^3.9"
click = "^7.1.2"
requests = "^2.25.1"
loguru = "^0.5.3"
"ruamel.yaml" = "^0.17.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
mypy = "^0.812"

[tool.poetry.scripts]
convo = 'convo.console:cli'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
